--- src/backend/expungeservice/test/test_file.py
@@ -1,3 +1,0 @@
-def test_run():
-  five = 5
-  assert five == 5
