+++ src/backend/tests/functional_tests/test_crawler_expunger.py
@@ -2,7 +2,7 @@ import unittest
 
 from datetime import date as date_class
 from dateutil.relativedelta import relativedelta
-from expungeservice.expungement_analyzer.expunger import Expunger
+from expungeservice.expunger.expunger import Expunger
 from tests.factories.crawler_factory import CrawlerFactory
 from tests.fixtures.case_details import CaseDetails
 from tests.fixtures.john_doe import JohnDoe
